パターン関係の引数が対応してなく型変換でオーバーフロウしていたので修正

+++ NNRiversi/Hive_routine.cpp
@@ -63,8 +63,8 @@ int NextMove(Hive *hive, const BitBoard *bitboard, char i_color, uint64 *PutPos)
 		else {
 			color = i_color;
 		}
-		return ((double)MidAlphaBeta(hive, hive->bitboard->stone[color], hive->bitboard->stone[oppColor(color)], -VALUE_MAX, VALUE_MAX,
-			FALSE, color, left, hive->midDepth, PutPos)+0.5)/STONE_VALUE;
+		return (int)(((double)MidAlphaBeta(hive, hive->bitboard->stone[color], hive->bitboard->stone[oppColor(color)], -VALUE_MAX, VALUE_MAX,
+			FALSE, color, left, hive->midDepth, PutPos)+0.5)/STONE_VALUE);
 	}
 }
